+++ frontend/app/src/main/java/com/example/vishingguard/smishing/SmishingViewModel.kt
@@ -1,8 +1,9 @@
-package com.example.vishingguard
+package com.example.vishingguard.smishing
 
 import android.util.Log
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import com.example.vishingguard.ServicePool
 import retrofit2.Call
 import retrofit2.Response
